Added index route handler to posts service, moved test communication to S3 to the route handler and exported AWS object with credentials and region from main index.ts server module.
